fixed the way the presence of config file is detected

+++ dspp_reader/tools/site.py
@@ -21,7 +21,7 @@ class Site(object):
             timezone=tz(self.timezone),
             description=self.name)
 
-    def get_time_range(self, sun_altitude:int = -10):
+    def get_time_range(self, sun_altitude:float = -10):
         now = Time(datetime.datetime.now(datetime.UTC))
         # now = Time("2024-12-02 09:00:00")
         reference_time = now
